Replace ColorMap with Gradient
Recent changes in the mandelbrot.clj have been using a vector of colors as opposed to the ColorMap. To make this easier, generate a Gradient record that will make the colors in between the given vector of colors.
This also allows for a default color to be given for the inner portion of the mandelbrot set