Move with-profiling
With profiling requires enabling certain settings on Linux prior to running the jar. Not enabling them causes a nasty stack-trace that is not obvious to where the error comes from.
Because of this, move the with-profiling macro into its namespace so that it doesn't get loaded at the same time as the utils namespace, and can be used as-needed