Allow for Parallel Generation to be Configured
The program has the ability to generate the mandelbrot values in parallel, or in-order. While this is possible, the value was hard-coded to always be parallel, because it is faster.
This options should be configurable by the user with the CLI flags. The default will still be true to speed up image generation.