Make a String to Color Function
Make a function that can take in a string, and return a java.awt.Color. If the string matches common colors (red, blue, green, etc.) use the pre-existing color for that. If it doesn't match, attempt to convert the value from hex. The leading 0x is stripped if it was given