Large Speed Improvement for Parallel
The Mandelbrot set is "embarrassingly parallelizable", but the previous approach only generated the values in parallel, but not the actual bytes in the final vector. This has been updated, which reduces the allocation of an intermediate double array. Each column number of bytes is now generated in parallel.
Testing locally, a 21000x21000 image takes roughly the same time as a 10000x10000 image in the previous parallel implementation.
Additionally, the required memory is MUCH lower