Use Log4J2 Logging With clojure.logging
Clojure has a log-wrapper, clojure.logging, but still requires that there be an underlying logger. Log4J2 isn't the simplest logger to use, but it's one I have experience with.

Because the JsonTemplateLayout extension is used when writing logs to a file, we need a custom Log4j2Plugins.dat resolver to merge in the uber jar. Inspiration for this is taken from: https://github.com/seancorfield/build-uber-log4j2-handler/tree/v2.23.1, with several changes to clean up the /temp directory, and to use static java.util.Collections functions